Add live Flow API tier (22 endpoints) + audit fixes
- Typed structs + untyped/*Typed client methods + godoc for all /v1/flow/*
  endpoints; reuse existing GexStrike/DexStrike for flow gex/dex
- Remove non-existent /v1/exposure/history endpoint
- Fix AdvTotalVarianceSurface.tenors []int -> []float64 (API returns
  fractional year-tenors)
- Per-endpoint live integration tests

+++ adv_volatility.go
@@ -98,8 +98,9 @@ type AdvTotalVarianceSurface struct {
 	Moneyness []float64 `json:"moneyness"`
 	// Expiries are the expiration dates corresponding to the inner axis.
 	Expiries []string `json:"expiries"`
-	// Tenors are the day-count tenors for each expiry (matches Expiries).
-	Tenors []int `json:"tenors"`
+	// Tenors are the year-fraction tenors for each expiry (matches
+	// Expiries) — fractional (e.g. 0.00274 ≈ 1 day), so float64.
+	Tenors []float64 `json:"tenors"`
 	// TotalVariance is the (moneyness × expiry) total-variance grid.
 	TotalVariance [][]float64 `json:"total_variance"`
 	// ImpliedVol is the (moneyness × expiry) implied-vol grid (annualised %).
